+++ x/oracle/client/cli/tx.go
@@ -24,6 +24,8 @@ func GetTxCmd() *cobra.Command {
 
 	cmd.AddCommand(
 		GetCmdDelegateFeedConsent(),
+		GetCmdAggregateExchangeRatePrevote(),
+		GetCmdAggregateExchangeRateVote(),
 	)
 
 	return cmd
@@ -52,6 +54,90 @@ func GetCmdDelegateFeedConsent() *cobra.Command {
 			}
 
 			msg := types.NewMsgDelegateFeedConsent(sdk.ValAddress(clientCtx.GetFromAddress()), feederAddr)
+
+			if err := msg.ValidateBasic(); err != nil {
+				return err
+			}
+
+			return tx.GenerateOrBroadcastTxCLI(clientCtx, cmd.Flags(), msg)
+		},
+	}
+
+	flags.AddTxFlagsToCmd(cmd)
+
+	return cmd
+}
+
+// GetCmdAggregateExchangeRatePrevote creates a Cobra command to generate or
+// broadcast a transaction with a MsgAggregateExchangeRatePrevote message.
+func GetCmdAggregateExchangeRatePrevote() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "exchange-rate-prevote [hash]",
+		Args:  cobra.ExactArgs(1),
+		Short: "Submit an exchange rate prevote with a hash",
+		Long: fmt.Sprintf(`Submit an exchange rate prevote with a hash as a hex string
+			representation of a byte array.
+			Ex: umeed tx oracle exchange-rate-prevote %s --from alice`,
+			"19c30cf9ea8aa0e0b03904162cadec0f2024a76d"),
+		RunE: func(cmd *cobra.Command, args []string) error {
+			clientCtx, err := client.GetClientTxContext(cmd)
+			if err != nil {
+				return err
+			}
+
+			hash, err := types.AggregateVoteHashFromHexString(args[0])
+			if err != nil {
+				return err
+			}
+
+			msg := types.NewMsgAggregateExchangeRatePrevote(
+				hash,
+				clientCtx.GetFromAddress(),
+				sdk.ValAddress(clientCtx.GetFromAddress()),
+			)
+
+			if err := msg.ValidateBasic(); err != nil {
+				return err
+			}
+
+			return tx.GenerateOrBroadcastTxCLI(clientCtx, cmd.Flags(), msg)
+		},
+	}
+
+	flags.AddTxFlagsToCmd(cmd)
+
+	return cmd
+}
+
+// GetCmdAggregateExchangeRateVote creates a Cobra command to generate or
+// broadcast a transaction with a NewMsgAggregateExchangeRateVote message.
+func GetCmdAggregateExchangeRateVote() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "exchange-rate-vote [salt] [exchange-rates]",
+		Args:  cobra.ExactArgs(2),
+		Short: "Submit an exchange rate vote with the salt and exchange rate string",
+		Long: fmt.Sprintf(`Submit an exchange rate vote with the salt of the previous hash, and the
+			exchange rate string previously used in the hash.
+			Ex: umeed tx oracle exchange-rate-vote %s %s --from alice`,
+			"0cf33fb528b388660c3a42c3f3250e983395290b75fef255050fb5bc48a6025f",
+			"foo:1.0,bar:1232.123",
+		),
+		RunE: func(cmd *cobra.Command, args []string) error {
+			clientCtx, err := client.GetClientTxContext(cmd)
+			if err != nil {
+				return err
+			}
+
+			msg := types.NewMsgAggregateExchangeRateVote(
+				args[0],
+				args[1],
+				clientCtx.GetFromAddress(),
+				sdk.ValAddress(clientCtx.GetFromAddress()),
+			)
+
+			if err := msg.ValidateBasic(); err != nil {
+				return err
+			}
 
 			return tx.GenerateOrBroadcastTxCLI(clientCtx, cmd.Flags(), msg)
 		},